Moved method up in the API layer
