+++ source/app/blueprints/graphql/permissions.py
@@ -22,10 +22,9 @@ from uuid import uuid4
 from flask import session
 from flask import request
 
-from app.blueprints.access_controls import get_case_access_from_api
+from app.blueprints.access_controls import get_case_access_from_api, ac_fast_check_current_user_has_case_access
 from app.blueprints.iris_user import iris_current_user
 from app.iris_engine.access_control.utils import ac_get_effective_permissions_of_user
-from app.business.access_controls import ac_fast_check_current_user_has_case_access
 
 
 class PermissionDeniedError(Exception):
