+++ source/app/blueprints/rest/v2/assets_routes/comments.py
@@ -20,7 +20,7 @@ from flask import Blueprint
 from flask import request
 from marshmallow.exceptions import ValidationError
 
-from app.blueprints.access_controls import ac_api_requires
+from app.blueprints.access_controls import ac_api_requires, ac_fast_check_current_user_has_case_access
 from app.blueprints.rest.endpoints import response_api_paginated
 from app.blueprints.rest.endpoints import response_api_not_found
 from app.blueprints.rest.endpoints import response_api_created
@@ -37,7 +37,6 @@ from app.blueprints.rest.case_comments import case_comment_update
 from app.business.assets import assets_get
 from app.business.errors import ObjectNotFoundError
 from app.schema.marshables import CommentSchema
-from app.business.access_controls import ac_fast_check_current_user_has_case_access
 from app.blueprints.iris_user import iris_current_user
 from app.models.authorization import CaseAccessLevel
