+++ source/app/blueprints/rest/v2/case_routes/events.py
@@ -20,7 +20,7 @@ from flask import Blueprint
 from flask import request
 from marshmallow.exceptions import ValidationError
 
-from app.blueprints.access_controls import ac_api_requires
+from app.blueprints.access_controls import ac_api_requires, ac_fast_check_current_user_has_case_access
 from app.blueprints.rest.endpoints import response_api_created
 from app.blueprints.rest.endpoints import response_api_success
 from app.blueprints.rest.endpoints import response_api_deleted
@@ -36,7 +36,6 @@ from app.schema.marshables import EventSchema
 from app.business.errors import BusinessProcessingError
 from app.business.errors import ObjectNotFoundError
 from app.business.cases import cases_exists
-from app.business.access_controls import ac_fast_check_current_user_has_case_access
 from app.iris_engine.utils.collab import notify
 from app.models.authorization import CaseAccessLevel
 from app.iris_engine.module_handler.module_handler import call_deprecated_on_preload_modules_hook
