+++ source/app/blueprints/rest/v2/case_routes/iocs.py
@@ -21,7 +21,7 @@ from flask import request
 from marshmallow import ValidationError
 
 from app.logger import logger
-from app.blueprints.access_controls import ac_api_requires
+from app.blueprints.access_controls import ac_api_requires, ac_fast_check_current_user_has_case_access
 from app.blueprints.rest.endpoints import response_api_created
 from app.blueprints.rest.endpoints import response_api_deleted
 from app.blueprints.rest.endpoints import response_api_not_found
@@ -37,7 +37,6 @@ from app.business.iocs import iocs_get
 from app.business.iocs import iocs_delete
 from app.business.iocs import iocs_update
 from app.datamgmt.case.case_iocs_db import get_filtered_iocs
-from app.business.access_controls import ac_fast_check_current_user_has_case_access
 from app.models.authorization import CaseAccessLevel
 from app.schema.marshables import IocSchemaForAPIV2
 from app.blueprints.access_controls import ac_api_return_access_denied
