+++ source/app/blueprints/rest/v2/case_routes/notes.py
@@ -20,9 +20,8 @@ from marshmallow import ValidationError
 from flask import Blueprint
 from flask import request
 
-from app.blueprints.access_controls import ac_api_requires
+from app.blueprints.access_controls import ac_api_requires, ac_fast_check_current_user_has_case_access
 from app.blueprints.access_controls import ac_api_return_access_denied
-from app.business.access_controls import ac_fast_check_current_user_has_case_access
 from app.blueprints.rest.endpoints import response_api_created
 from app.blueprints.rest.endpoints import response_api_success
 from app.blueprints.rest.endpoints import response_api_deleted
