+++ source/app/blueprints/rest/v2/case_routes/tasks.py
@@ -27,7 +27,7 @@ from app.blueprints.rest.endpoints import response_api_success
 from app.blueprints.rest.endpoints import response_api_paginated
 from app.blueprints.rest.endpoints import response_api_created
 from app.blueprints.rest.parsing import parse_pagination_parameters
-from app.blueprints.access_controls import ac_api_return_access_denied
+from app.blueprints.access_controls import ac_api_return_access_denied, ac_fast_check_current_user_has_case_access
 from app.blueprints.access_controls import ac_api_requires
 from app.schema.marshables import CaseTaskSchema
 from app.business.errors import BusinessProcessingError
@@ -38,7 +38,6 @@ from app.business.tasks import tasks_update
 from app.business.tasks import tasks_delete
 from app.business.tasks import tasks_filter
 from app.models.authorization import CaseAccessLevel
-from app.business.access_controls import ac_fast_check_current_user_has_case_access
 from app.iris_engine.module_handler.module_handler import call_deprecated_on_preload_modules_hook
