+++ source/app/blueprints/rest/v2/cases.py
@@ -46,8 +46,7 @@ from app.business.cases import cases_update
 from app.business.errors import BusinessProcessingError
 from app.datamgmt.manage.manage_cases_db import get_filtered_cases
 from app.schema.marshables import CaseSchemaForAPIV2
-from app.blueprints.access_controls import ac_api_requires
-from app.business.access_controls import ac_fast_check_current_user_has_case_access
+from app.blueprints.access_controls import ac_api_requires, ac_fast_check_current_user_has_case_access
 from app.blueprints.access_controls import ac_api_return_access_denied
 from app.models.authorization import Permissions
 from app.models.authorization import CaseAccessLevel
